requirements.txt
# base-modules
asgiref
Django
sqlparse
tzdata

# generate fake-data [DEV]
faker

# image
pillow

# API
djangorestframework
markdown
django-filter
drf-yasg[validation]
djangorestframework-simplejwt
requests

# encrypt
cryptography

# third-party-apps
django-filter
django-cors-headers
django-cors-middleware
django-mail-templated
python-decouple

# websocket
channels[daphne]
channels-redis

# delivering_data
redis

# worker
celery
django-celery-beat

# finance-data -> cache
yfinance

# cache
django-redis

# test
black
flake8
